Refactor project structure to use multiple smaller crates
- Introduced a new workspace structure with separate crates for `ai-client`, `python-bridge`, and `translator`.
- Implemented `ai-client` crate with models for ChatGPT and DeepL, including API client and glossary handling (models).
- Added `python-bridge` crate for Python integration, enabling calls to Python scripts (old process_call).
- Updated `Cargo.toml` to reflect new dependencies and workspace members.
- Refactored handlers to utilize the new `ai-client` and `python-bridge` modules.
- Removed obsolete modules and files related to previous structure.
