+++ src-tauri/crates/ai-client/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "ai-client"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+calamine = "0.31.0"
+serde_json = "1"
+serde = { version = "1", features = ["derive"] }
+reqwest = "0.12.12"
