+++ src-tauri/crates/ai-client/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod structs;
+pub mod models;
+
+pub use structs::deepl::Deepl;
+pub use structs::chatgpt::ChatGPT;
