+++ src-tauri/crates/ai-client/src/models/mod.rs
@@ -1,0 +1,3 @@
+pub mod chatgpt;
+pub mod deepl;
+pub mod model;
