+++ src-tauri/crates/ai-client/src/structs/api_client.rs
@@ -1,0 +1,9 @@
+use reqwest::Client;
+use reqwest::header::HeaderMap;
+
+#[derive(Debug)]
+pub struct APIClient {
+  pub(crate) api_key: String,
+  pub client: Client,
+  pub headers: HeaderMap
+}
