+++ src-tauri/crates/ai-client/src/structs/chatgpt.rs
@@ -1,0 +1,5 @@
+use crate::structs::api_client::APIClient;
+
+pub struct ChatGPT {
+  pub client: APIClient,
+}
