+++ src-tauri/crates/ai-client/src/structs/deepl.rs
@@ -1,0 +1,5 @@
+use crate::structs::api_client::APIClient;
+
+pub struct Deepl {
+  pub client: APIClient,
+}
