+++ src-tauri/crates/ai-client/src/structs/dictionary.rs
