+++ src-tauri/crates/ai-client/src/structs/glossary.rs
