+++ src-tauri/crates/ai-client/src/structs/mod.rs
@@ -1,0 +1,6 @@
+pub mod language;
+pub mod dictionary;
+pub mod glossary;
+pub mod api_client;
+pub mod deepl;
+pub mod chatgpt;
