+++ src-tauri/crates/python-bridge/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "python-bridge"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+pyo3 = { version = "0.23.3", default-features = false }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
