+++ src-tauri/crates/translator/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "translator"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+calamine = "0.31.0"
+tiktoken-rs = "0.9.1"
+datetime = "0.5.2"
+docx-rs = "0.4.18"
+ooxmlsdk = "0.3.0"
