--- src-tauri/src/models/mod.rs
@@ -1,3 +1,0 @@
-pub mod model;
-pub mod chatgpt;
-pub mod deepl;
