--- src-tauri/src/structs/mod.rs
@@ -1,3 +1,0 @@
-pub mod language;
-pub mod dictionary;
-pub mod glossary;
