--- src-tauri/src/translator/mod.rs
@@ -1,1 +1,0 @@
-pub mod document;
